fix int length

+++ vsys/nft.go
@@ -21,7 +21,7 @@ type NFTIssueResponse struct {
 
 type NFTContract struct {
 	SenderAddress string
-	TokenIndex    int32
+	TokenIndex    int64
 	Recipient     string
 	ContractId    string
 	Description   string
@@ -70,7 +70,7 @@ func (c *NFTContract) DecodeSendData(data []byte) {
 	de := DataEncoder{}
 	list := de.Decode(data)
 	c.Recipient = list[0].Value.(string)
-	c.TokenIndex = list[1].Value.(int32)
+	c.TokenIndex = list[1].Value.(int64)
 }
 
 func (c *NFTContract) DecodeTransferData(data []byte) {
@@ -78,7 +78,7 @@ func (c *NFTContract) DecodeTransferData(data []byte) {
 	list := de.Decode(data)
 	c.SenderAddress = list[0].Value.(string)
 	c.Recipient = list[1].Value.(string)
-	c.TokenIndex = list[2].Value.(int32)
+	c.TokenIndex = list[2].Value.(int64)
 }
 
 func (c *NFTContract) DecodeDepositData(data []byte) {
@@ -86,7 +86,7 @@ func (c *NFTContract) DecodeDepositData(data []byte) {
 	list := de.Decode(data)
 	c.SenderAddress = list[0].Value.(string)
 	c.ContractId = list[1].Value.(string)
-	c.TokenIndex = list[2].Value.(int32)
+	c.TokenIndex = list[2].Value.(int64)
 }
 
 func (c *NFTContract) DecodeWithdrawData(data []byte) {
@@ -94,7 +94,7 @@ func (c *NFTContract) DecodeWithdrawData(data []byte) {
 	list := de.Decode(data)
 	c.ContractId = list[0].Value.(string)
 	c.Recipient = list[1].Value.(string)
-	c.TokenIndex = list[2].Value.(int32)
+	c.TokenIndex = list[2].Value.(int64)
 }
 
 func (c *NFTContract) BuildRegisterData() []byte {
